divide main.cpp into multiple files
