+++ Arena.hpp
@@ -1,0 +1,3 @@
+#pragma once
+#include"Pokemon.hpp"
+void turnperturn(Pokemon& pokemon1, Pokemon& pokemon2);
