+++ PokemonTypeEnum.h
@@ -1,0 +1,2 @@
+#pragma once
+enum class pokemontype { fire, water, grass, fly, ground };
